🏗️ Replace solution for day 5, 2016 with a Rust/WASM implementation
This almost cuts the runtime for this day in half, from 12-14 secs to 7
secs. Not as much as I would have expected but only about twice as slow
as the same Rust program running natively in release at about 3 secs. I
guess the real takeaway is that the JS version is not orders of
magnitude slower.

This commit also lays the foundation for solving future days in
Rust/WASM when we get to the hard optimization problems.