🧰 Remove naughty `Object.prototype` injection in favor of a plain old function
While the `.iter()` was syntactically much nicer, injecting into foreign
prototypes is by all accounts a naughty thing and we don't want to be
naughty, so Santa brings us presents.

In all honesty, I spend the afternoon debugging a weird glitch in the
`rimbu/hashed` package which ultimately turned out to be caused by an
`iter()` method buried deep in the implementation which clashed with
the prototype. Yikes. Lesson learned, I guess.